Changes in cqcHeader: Changed a typo in the __init__ method for CQCNotifyHeader. The change is self.unpack(self,headerBytes) -> self.unpack(headerBytes)
Changes in cqcProtocol:
Added an argument called length in _send_back_cqc which defaults to 0. This sets the length of the additional data in the CQCHeader.
The following methods now call _send_back_cqc with length=CQC_NOTIFY_LENGTH
* handle_time
* cmd_measure
* cmd_recv

Changes in pythonLib/cqc.py:
Small change in rececie, just for clearity